rename the receiver module
