--- qemu/overlay.normal/etc/zinit/api-receiver.yaml
@@ -1,1 +1,0 @@
-../../../../etc/zinit/api-receiver.yaml
