+++ qemu/overlay.normal/etc/zinit/myceliumrx.yaml
@@ -1,0 +1,1 @@
+../../../../etc/zinit/myceliumrx.yaml
